success on postgres

+++ database/migrations/20180407063726-create-users-table.js
@@ -37,7 +37,7 @@ module.exports = {
       allowNull: false,
     },
     status: {
-      type: Sequelize.TINYINT,
+      type: Sequelize.INTEGER,
       defaultValue: 1,
       allowNull: false,
     },
